refactor: split the toast channel out of the toast provider
`toast.tsx` exported both `LibraryToastProvider` and `useLibraryToast`, so
`react-refresh/only-export-components` fired on it: a module that exports a
component alongside a plain function can't be hot-swapped, and an edit to the
toast markup would full-reload the app rather than Fast Refresh.

`ToastContext`, `ShowToast`, `ToastTone` and `useLibraryToast` move to
`state/toast.context.ts` — the same provider/context split the codebase
already uses for `open-change-requests`, `git`, `layout` and the rest of the
`*.context.ts` modules. `toast.tsx` keeps the provider and `TONE`, which is a
rendering detail of the markup that consumes it.

Behaviour is unchanged: the six `useLibraryToast` call sites import from the
new path, and `LibraryToastProvider` stays where it was, so tests and
`LibraryRoutes` are untouched.
