+++ packages/core-frontend/src/modules/library/components/NewGroupDialog.tsx
@@ -6,7 +6,7 @@ import { createDirectory } from '../../workspace/services/workspace.api';
 import { GROUPS_DIR } from '@bevel-software/platform-shared';
 import { DEFAULT_WORKSPACE_ID } from '../services/library.api';
 import { pathForGroup } from '../routes/library-paths';
-import { useLibraryToast } from '../state/toast';
+import { useLibraryToast } from '../state/toast.context';
 
 export interface NewGroupDialogProps {
   /** Names already taken — readable groups AND locked ones. */
